+++ app/src/main/java/com/sogukj/pe/ui/news/MainNewsFragment.kt
@@ -281,7 +281,7 @@ class MainNewsFragment : BaseFragment() {
                 .subscribe({ payload ->
                     if (payload.isOk) {
                         payload?.apply {
-                            tv_result_title.text = Html.fromHtml(getString(R.string.tv_title_result_news, total))
+                            tv_result_title.text = Html.fromHtml(getString(R.string.tv_title_result_news, (total as Double).toInt()))
                         }
                         if (page == 1)
                             adapter.dataList.clear()
